투표 앱과 투표 결과 앱 분리
