+++ css/movie-item.css
@@ -1,0 +1,39 @@
+.item {
+	list-style: none;
+	padding:0px 5px;
+	margin:5px;
+	border:1px solid #ddd;
+}
+
+.item {
+	line-height: 14px;
+	cursor:pointer;
+}
+
+.item:hover {
+	background-color:#aaa;
+	color:#fff;
+}
+
+.itemBasic {
+	padding:10px;
+	vertical-align: middle;
+}
+
+.itemMO {
+	display:none;
+	border-top:1px solid #efefef;
+	padding:10px;
+}
+
+.item:hover > .itemMO {
+	display:block;
+}
+
+.director_name {
+	font-size:.7em;
+}
+
+.premier {
+	font-size:.7em;
+}
