+++ css/result-movie-item.css
@@ -1,0 +1,35 @@
+.item,
+.itemFinished,
+.itemSelected {
+	list-style: none;
+	padding:0px 5px;
+	margin:5px;
+	border:1px solid #ddd;
+}
+
+.itemSelected {
+	background-color:#ffdddd;
+	background-size:50%;
+}
+.itemBasic {
+	padding:10px;
+	vertical-align: middle;
+}
+
+.voted {
+	font-size:.7em;
+	color:#ff0000;
+	border:1px solid #ff0000;
+	padding:1px 2px;
+}
+.director_name {
+	font-size:.7em;
+}
+
+.premier {
+	font-size:.7em;
+}
+
+.rate {
+	float:right;
+}
