+++ css/vote.css
@@ -11,60 +11,6 @@
 	padding:0;
 }
 
-.item,
-.itemFinished,
-.itemSelected {
-	list-style: none;
-	padding:0px 5px;
-	margin:5px;
-	border:1px solid #ddd;
-}
-
-.item {
-	line-height: 14px;
-	cursor:pointer;
-}
-.itemSelected {
-	background-color:#ffdddd;
-	background-size:50%;
-}
-.item:hover {
-	background-color:#aaa;
-	color:#fff;
-}
-
-.itemBasic {
-	padding:10px;
-	vertical-align: middle;
-}
-.itemMO {
-	display:none;
-	border-top:1px solid #efefef;
-	padding:10px;
-}
-
-.item:hover > .itemMO {
-	display:block;
-}
-
-.voted {
-	font-size:.7em;
-	color:#ff0000;
-	border:1px solid #ff0000;
-	padding:1px 2px;
-}
-.director_name {
-	font-size:.7em;
-}
-
-.premier {
-	font-size:.7em;
-}
-
-.rate {
-	float:right;
-}
-
 .searchApp {
 	text-align:center;
 	padding:5px;
